[feat] /home 라우트 추가
